Fix assault troops from being used by defensive AI. fix too many starbases and not enough star bases. rework spy economy to allow them to be useful late game with crushing the economy early game. wired in a separate spy budget, which works way better,  but not implemented due to incompatibility with previous saves. rework ssp budget in a like way. ssps still need work. seeing orphaned ssp's.
--HG--
branch : feature/Push here instead of develop

+++ Ship_Game/Empire.cs
@@ -2299,7 +2299,7 @@ namespace Ship_Game
                 this.data.AgentList[index].ServiceYears += 0.1f;
             }
             this.data.AgentList.ApplyPendingRemovals();
-            if ((double)this.Money < 0.0)
+            if (this.Money < 0.0)
             {
                 ++this.data.TurnsBelowZero;
             }
